Modify template for firewall rule for when dns is automatically configured
Caddy will automatically receive requests from port 443 when DNS is configured and
the Caddyfile has a domain name in it.
Right now it is only listening to port 80 and will fail on https call. I tried to
change the default Caddyfile port to 443 but it seems I may need to do more wrangling
to get that to work as expected.